ルーティング名整理

+++ routes/web.php
@@ -23,9 +23,17 @@ Route::get('/', function () {
     return view('auth/login');
 });
 
-Route::get('/dashboard', function () {
-    return view('dashboard');
-})->middleware(['auth'])->name('dashboard');
+Route::get('/', function () {
+    if (auth()->check()) {
+        if (auth()->user()->can('isDeveloper')) {
+            return redirect()->route('users.index');
+        } elseif (auth()->user()->can('isAdmin')) {
+            return redirect()->route('admin.reservations.reserved_index');
+        } elseif (auth()->user()->can('isUser')) {
+            return redirect()->route('user.reservations.reservable_index');
+        }
+    }
+})->middleware(['auth']);
 
 // 開発者用
 Route::middleware(['auth', 'can:isDeveloper'])->group(function () {
@@ -56,43 +64,43 @@ Route::middleware(['auth', 'can:isDeveloper'])->group(function () {
 // 一般管理者用
 Route::middleware(['auth', 'can:isAdmin'])->group(function () {
     //予約管理、予約可能日一覧、予約済み一覧、登録、編集、キャンセル
-    Route::prefix('reservations')->group(function(){
-        Route::get('', [AdminReservationController::class, 'index'])->name('reservations.index');
-        Route::get('/reservable_index', [AdminReservationController::class, 'reservableIndex'])->name('reservations.reservable_index');
-        Route::get('create', [AdminReservationController::class, 'create'])->name('reservations.create');
-        Route::post('create', [AdminReservationController::class, 'store'])->name('reservations.store');
-        Route::get('{id}/edit', [AdminReservationController::class, 'edit'])->name('reservations.edit');
-        Route::post('{id}/edit', [AdminReservationController::class, 'update'])->name('reservations.update');
-        Route::post('{reservation_id}/cancel', [AdminReservationController::class, 'cancel'])->name('reservations.cancel');
-        Route::delete('{id}', [AdminReservationController::class, 'delete'])->name('reservations.delete');
+    Route::prefix('admin/reservations')->group(function(){
+        Route::get('/reserved_index', [AdminReservationController::class, 'reservedIndex'])->name('admin.reservations.reserved_index');
+        Route::get('/reservable_index', [AdminReservationController::class, 'reservableIndex'])->name('admin.reservations.reservable_index');
+        Route::get('/create', [AdminReservationController::class, 'create'])->name('admin.reservations.create');
+        Route::post('/create', [AdminReservationController::class, 'store'])->name('admin.reservations.store');
+        Route::get('{id}/edit', [AdminReservationController::class, 'edit'])->name('admin.reservations.edit');
+        Route::post('{id}/edit', [AdminReservationController::class, 'update'])->name('admin.reservations.update');
+        Route::post('{reservation_id}/cancel', [AdminReservationController::class, 'cancel'])->name('admin.reservations.cancel');
+        Route::delete('{id}', [AdminReservationController::class, 'delete'])->name('admin.reservations.delete');
     });
     
     // ユーザー情報、照会、編集
     Route::prefix('admin')->group(function(){
-        Route::get('', [AuthController::class, 'show'])->name('auth.show');
-        Route::get('/edit', [AuthController::class, 'edit'])->name('auth.edit');
-        Route::post('/update', [AuthController::class, 'update'])->name('auth.update');
+        Route::get('', [AuthController::class, 'show'])->name('admin.show');
+        Route::get('/edit', [AuthController::class, 'edit'])->name('admin.edit');
+        Route::post('/update', [AuthController::class, 'update'])->name('admin.update');
     });
 });
 
 // 一般ユーザー用
 Route::middleware(['auth', 'can:isUser'])->group(function () {
     // 予約可能一覧、自分の予約一覧、予約登録、予約確認、キャンセル
-    Route::prefix('reservables')->group(function(){
-        Route::get('', [UserReservationController::class, 'index'])->name('reservables.index');
-        Route::get('/reserved_index', [UserReservationController::class, 'reservedIndex'])->name('reservables.resereved_index');
-        Route::get('{reservation_id}/edit', [UserReservationController::class, 'edit'])->name('reservables.edit');
-        Route::post('{reservation_id}/regist', [UserReservationController::class, 'regist'])->name('reservables.regist');
-        Route::get('{reservation_id}/show', [UserReservationController::class, 'show'])->name('reservables.show');
-        Route::post('{reservation_id}/cancel', [UserReservationController::class, 'cancel'])->name('reservables.cancel');
+    Route::prefix('user/reservations')->group(function(){
+        Route::get('/reservable_index', [UserReservationController::class, 'reservableIndex'])->name('user.reservations.reservable_index');
+        Route::get('/reserved_index', [UserReservationController::class, 'reservedIndex'])->name('user.reservations.resereved_index');
+        Route::get('{reservation_id}/edit', [UserReservationController::class, 'edit'])->name('user.reservations.edit');
+        Route::post('{reservation_id}/regist', [UserReservationController::class, 'regist'])->name('user.reservations.regist');
+        Route::get('{reservation_id}/show', [UserReservationController::class, 'show'])->name('user.reservations.show');
+        Route::post('{reservation_id}/cancel', [UserReservationController::class, 'cancel'])->name('user.reservations.cancel');
     });
     
     // ユーザー情報、照会、編集、退会
-    Route::prefix('auth')->group(function(){
-        Route::get('', [AuthController::class, 'show'])->name('auth.show');
-        Route::get('/edit', [AuthController::class, 'edit'])->name('auth.edit');
-        Route::post('/update', [AuthController::class, 'update'])->name('auth.update');
-        Route::delete('delete', [AuthController::class, 'delete'])->name('auth.delete');
+    Route::prefix('user')->group(function(){
+        Route::get('', [AuthController::class, 'show'])->name('user.show');
+        Route::get('/edit', [AuthController::class, 'edit'])->name('user.edit');
+        Route::post('/update', [AuthController::class, 'update'])->name('user.update');
+        Route::delete('delete', [AuthController::class, 'delete'])->name('user.delete');
     });
 });
